Update install.sh
Color Cleanup and not on future script needed and where to seperate this into that script since "arch-chroot" breaks scripts and run that scrip by running it as a second argument on "arch-chroot" or have user do that after they get the new shell instance. New script may need to be ran from new root user home.