Update install.sh
Replacing ARGS with a whiptail to allow some change in the at runtime and use a second file to do all the work and allow ARGS to allow automation. how the data is exchanged has not been determined at this time but when that happens appropriate changes will take place.